drm/msm/dsi-staging: update dsi secure mode flag properly
In the current implementation dsi secure flag is set to false
during probe. But after that this flag is never updated even
though secure domain attach/detach is handled in dsi driver.
So when secure session is established and the non-secure buffers
are unmapped, since the flag is not updated dsi driver is still
trying to access the now invalid non-secure buffers and this is
causing kernel panic. Fix this issue by updating the dsi secure
flag properly during the secure session callbacks.

Change-Id: I10801aacb4cd1925472f58331fa2f5a56eff6aa6
Signed-off-by: Sandeep Panda <spanda@codeaurora.org>